requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
Pillow==10.4.0
qrcode==7.4.2
reportlab==4.2.2
itsdangerous==2.2.0
Werkzeug==3.0.4
email-validator==2.2.0
